want to render but render crashed

+++ includes/components/camera.h
@@ -6,7 +6,7 @@ enum class CameraType {
 };
 
 struct Camera {
-  double fov;
-  double near;
-  double far;
+  float fov;
+  float near;
+  float far;
 };
